BE-50, refactor: move UserBung related test code
